requirements.txt
python-telegram-bot==13.12
requests
flask
